requirements.txt
fastapi
uvicorn
scikit-learn
pandas
nltk
openai
python-dotenv
